# Speed up by switching to hstream