--- a/package.json
+++ b/package.json
@@ -14,7 +14,7 @@
     "extract-html-class": "^1.0.1",
     "extract-html-id": "^1.0.0",
     "extract-html-tag": "^1.0.1",
-    "hyperstream": "^1.2.2",
+    "hstream": "^1.0.0",
     "through2": "^2.0.3"
   },
   "devDependencies": {