[hstream](https://github.com/goto-bus-stop/hstream) is a fast
hyperstream alternative. It lacks some features, but inline-critical-css
only needs to append some tags to the `<head>`, and that _does_ work.

Unscientific test suggests ~2× speedup:

```
$ git checkout master
$ time node test
...
node test  0.27s user 0.02s system 105% cpu 0.269 total
```

```
$ git checkout hstream
$ time node test
...
node test  0.14s user 0.01s system 104% cpu 0.147 total
```

This should help bankai SSR be a bit speedier, combined with the work in
https://github.com/choojs/bankai/pull/394