@yoshuawuyts 2018-01-24
`v1.2.0` :tada: